requirements.txt
click>=8.1
httpx>=0.27
gitpython>=3.1
pyyaml>=6.0
prometheus-client>=0.20
